requirements.txt
Flask==3.0.3
mysql-connector-python==8.4.0
PyPDF2==3.0.1
scikit-learn==1.5.0
numpy==1.26.4
python-dotenv==1.0.1
